backend/requirements.txt
# Core Web Framework
fastapi==0.128.0
starlette==0.50.0
uvicorn==0.40.0
gunicorn==25.0.3
python-multipart==0.0.22
sse-starlette==3.2.0

# Database
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
alembic==1.18.1
pgvector==0.4.2

# Validation & Serialization
pydantic==2.12.5
pydantic-settings==2.12.0
email_validator==2.3.0
orjson==3.11.7

# Authentication & Security
argon2-cffi>=25.1.0
PyJWT[crypto]==2.11.0
cryptography==46.0.5
pyotp==2.9.0
qrcode==8.2

# Async & Background Tasks
celery==5.6.2
redis==6.4.0
broadcaster[redis]==0.3.1
anyio==4.12.1

# HTTP Client
httpx==0.28.1
requests==2.32.5

# Email
resend==2.21.0
Jinja2==3.1.6
pywebpush>=2.0.0
twilio>=8.0.0

# Search & NLP
symspellpy==6.9.0
dateparser>=1.2.0
openai==2.18.0

# Geospatial
geopandas==1.1.2

# Payments
stripe==14.3.0

# Monitoring & Utilities
prometheus-client==0.24.1
sentry-sdk[fastapi]==2.52.0
# OpenTelemetry Core - PIN TO SAME MINOR VERSION
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
# OTLP Exporter
opentelemetry-exporter-otlp-proto-http==1.39.1
# Auto-instrumentation - ALL SAME MINOR (0.43b0)
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-celery==0.60b1
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-instrumentation-logging==0.60b1
psutil==7.2.1
python-ulid==3.1.0
python-dotenv==1.2.1
PyYAML==6.0.3
user-agents==2.2.0
pillow==12.1.1
filelock>=3.20.3
urllib3>=2.6.3
virtualenv>=20.36.1
werkzeug>=3.1.5

# CLI Tools (for scripts)
click==8.3.1
colorama==0.4.6

# Development/Testing (could move to requirements-dev.txt)
pytest==9.0.2
pytest-asyncio==1.3.0

# Dev script dependencies (only used in scripts/dev/)
matplotlib==3.10.8
networkx==3.6.1
hiredis==3.3.0
numpy==2.4.1
# pytest-timeout moved to requirements-dev.txt (testing dependency)
